@@ -4,7 +4,7 @@
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>2.6.7</version>
+        <version>3.0.5</version>
 		<relativePath /> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>com.movieEureka</groupId>
@@ -13,9 +13,9 @@
 	<name>movieEureka</name>
 	<description>movie eureka server</description>
 	<properties>
-		<java.version>11</java.version>
-		<kotlin.version>1.6.0</kotlin.version>
-		<spring-cloud.version>2021.0.0</spring-cloud.version>
+		<java.version>17</java.version>
+		<kotlin.version>1.7.22</kotlin.version>
+		<spring-cloud.version>2022.0.2</spring-cloud.version>
 	</properties>
 	<dependencies>
 		<dependency>
