@@ -4,7 +4,7 @@
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>2.6.7</version>
+		<version>3.0.5</version>
 		<relativePath /> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>com.movieSchedule</groupId>
@@ -13,11 +13,11 @@
 	<name>movieSchedule</name>
 	<description>microservice Spring Lab</description>
 	<properties>
-		<java.version>11</java.version>
-		<kotlin.version>1.6.0</kotlin.version>
-		<testcontainers.version>1.16.0</testcontainers.version>
-		<springfox.version>3.0.0</springfox.version>
-		<spring-cloud.version>2021.0.0</spring-cloud.version>
+		<java.version>17</java.version>
+		<kotlin.version>1.7.22</kotlin.version>
+		<spring-cloud.version>2022.0.2</spring-cloud.version>
+		<testcontainers.version>1.17.6</testcontainers.version>
+		<springdoc.version>2.1.0</springdoc.version>
 	</properties>
 	<dependencies>
 		<dependency>
@@ -37,14 +37,9 @@
 			<artifactId>postgresql</artifactId>
 		</dependency>
 		<dependency>
-			<groupId>io.springfox</groupId>
-			<artifactId>springfox-boot-starter</artifactId>
-			<version>${springfox.version}</version>
-		</dependency>
-		<dependency>
-			<groupId>io.springfox</groupId>
-			<artifactId>springfox-swagger-ui</artifactId>
-			<version>${springfox.version}</version>
+			<groupId>org.springdoc</groupId>
+			<artifactId>springdoc-openapi-starter-webmvc-ui</artifactId>
+			<version>${springdoc.version}</version>
 		</dependency>
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
