@@ -1,10 +1,10 @@
 package com.movieSchedule.movieSchedule.domain
 
-import javax.persistence.GeneratedValue
-import javax.persistence.Id
-import javax.persistence.Column
+import jakarta.persistence.Column
+import jakarta.persistence.Entity
+import jakarta.persistence.GeneratedValue
+import jakarta.persistence.Id
 import java.time.LocalDate
-import javax.persistence.Entity
 
 @Entity
 class MovieOrder(
