@@ -1,14 +1,12 @@
 package com.movieShop.domain
 
 import com.movieShop.dto.MovieDto
-import javax.persistence.Column
-import javax.persistence.Entity
-import javax.persistence.GeneratedValue
-import javax.persistence.Id
-import javax.persistence.JoinColumn
-import javax.persistence.ManyToOne
-import org.hibernate.annotations.Fetch
-import org.hibernate.annotations.FetchMode
+import jakarta.persistence.Column
+import jakarta.persistence.Entity
+import jakarta.persistence.GeneratedValue
+import jakarta.persistence.Id
+import jakarta.persistence.JoinColumn
+import jakarta.persistence.ManyToOne
 
 @Entity
 class Movie(
