@@ -1,9 +1,9 @@
 package com.movieShop.dto
 
 import com.movieShop.domain.Producer
-import javax.validation.constraints.NotBlank
-import javax.validation.constraints.NotEmpty
-import javax.validation.constraints.NotNull
+import jakarta.validation.constraints.NotBlank
+import jakarta.validation.constraints.NotEmpty
+import jakarta.validation.constraints.NotNull
 
 data class ProducerDto(
         var id: Long? = null,
